moved tracer to a separate file
